refactor(contract): move the wire client and stubs to the contract side
Both sides of the boundary need the wire pieces: the daemon serves the
protocol, the agent dials it. They therefore belong to the contract
package, not to the supervisor implementation:

- aleph.vm.supervisor._pb        -> aleph.vm.supervisor_interface.wire._pb
- aleph.vm.supervisor.proto_convert
                                 -> aleph.vm.supervisor_interface.wire.proto_convert
- aleph.vm.supervisor.grpc_client
                                 -> aleph.vm.supervisor_interface.client
- ERROR_TRAILER_KEY moves from grpc_server to supervisor_interface.wire
  (it is a wire-protocol constant shared by client and server)

With this, the agent's production path no longer imports any
aleph.vm.supervisor module: the import-linter exemption for
agent.supervisor -> supervisor.grpc_client is gone. generate_proto.py,
check_proto_clean.sh and the tool excludes (ruff/isort/coverage/mypy)
follow the new _pb location; the generated bindings are unchanged
(the _pb2_grpc module already used a relative import).
