+++ scripts/check_proto_clean.sh
@@ -18,16 +18,16 @@ cd "$(dirname "$0")/.."
 python scripts/generate_proto.py
 
 if ! git diff --quiet --exit-code -- \
-    src/aleph/vm/supervisor/_pb/supervisor_pb2.py \
-    src/aleph/vm/supervisor/_pb/supervisor_pb2_grpc.py \
+    src/aleph/vm/supervisor_interface/wire/_pb/supervisor_pb2.py \
+    src/aleph/vm/supervisor_interface/wire/_pb/supervisor_pb2_grpc.py \
     proto/; then
     echo
     echo "ERROR: generated proto bindings are out of date." >&2
     echo "Run: python scripts/generate_proto.py" >&2
     echo "Then commit the changes." >&2
     git diff -- \
-        src/aleph/vm/supervisor/_pb/supervisor_pb2.py \
-        src/aleph/vm/supervisor/_pb/supervisor_pb2_grpc.py \
+        src/aleph/vm/supervisor_interface/wire/_pb/supervisor_pb2.py \
+        src/aleph/vm/supervisor_interface/wire/_pb/supervisor_pb2_grpc.py \
         proto/
     exit 1
 fi
