+++ tests/integration/test_daemon_restart.py
@@ -14,7 +14,7 @@ from conftest import (
     wait_for_tcp_banner,
 )
 
-from aleph.vm.supervisor.grpc_client import GrpcSupervisor
+from aleph.vm.supervisor_interface.client import GrpcSupervisor
 from aleph.vm.supervisor_interface.types import VmStatus
 
 pytestmark = pytest.mark.asyncio
